+++ RbacConsoleModule.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace dektrium\rbac;
+
+use yii\base\Module;
+
+/**
+ * Rbac console module.
+ * 
+ * @author Dmitry Erofeev <dmeroff@gmail.com>
+ */
+class RbacConsoleModule extends Module
+{
+    /**
+     * @inheritdoc
+     */
+    public $controllerNamespace = 'dektrium\rbac\commands';
+}