+++ commands/MigrateController.php
@@ -1,0 +1,128 @@
+<?php
+
+/*
+ * This file is part of the Dektrium project.
+ *
+ * (c) Dektrium project <http://github.com/dektrium/>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+
+namespace dektrium\rbac\commands;
+
+use yii\console\controllers\BaseMigrateController;
+use yii\db\Connection;
+use yii\db\Query;
+use yii\di\Instance;
+use yii\helpers\Console;
+
+/**
+ * Manages rbac migrations.
+ *
+ * @author Dmitry Erofeev <dmeroff@gmail.com>
+ */
+class MigrateController extends BaseMigrateController
+{
+    /**
+     * @var string the name of the table for keeping applied migration information.
+     */
+    public $migrationTable = '{{%auth_migration}}';
+
+    /**
+     * @var string the directory storing the migration classes. This can be either
+     * a path alias or a directory.
+     */
+    public $migrationPath = '@app/rbac/migrations';
+
+    /**
+     * @var Connection|array|string the DB connection object or the application component ID of the DB connection to use
+     * when applying migrations.This can also be a configuration array for creating the object.
+     */
+    public $db = 'db';
+
+    /**
+     * @inheritdoc
+     */
+    public $templateFile = '@dektrium/rbac/views/migration.php';
+
+    /**
+     * @inheritdoc
+     */
+    public function init()
+    {
+        parent::init();
+
+        $this->db = Instance::ensure($this->db, Connection::className());
+    }
+
+    /**
+     * @return array|string|Connection
+     */
+    public function getDb()
+    {
+        return $this->db;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function getMigrationHistory($limit)
+    {
+        if ($this->db->schema->getTableSchema($this->migrationTable, true) === null) {
+            $this->createMigrationHistoryTable();
+        }
+
+        $history = (new Query())
+            ->select(['apply_time'])
+            ->from($this->migrationTable)
+            ->orderBy(['apply_time' => SORT_DESC, 'version' => SORT_DESC])
+            ->limit($limit)
+            ->indexBy('version')
+            ->column($this->db);
+
+        unset($history[self::BASE_MIGRATION]);
+
+        return $history;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function addMigrationHistory($version)
+    {
+        $this->db->createCommand()->insert($this->migrationTable, [
+            'version'    => $version,
+            'apply_time' => time(),
+        ])->execute();
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function removeMigrationHistory($version)
+    {
+        $this->db->createCommand()->delete($this->migrationTable, [
+            'version' => $version,
+        ])->execute();
+    }
+
+    /**
+     * Creates the migration history table.
+     */
+    protected function createMigrationHistoryTable()
+    {
+        $tableName = $this->db->schema->getRawTableName($this->migrationTable);
+        $this->stdout("Creating migration history table \"$tableName\"...", Console::FG_YELLOW);
+        $this->db->createCommand()->createTable($this->migrationTable, [
+            'version'    => 'VARCHAR(180) NOT NULL PRIMARY KEY',
+            'apply_time' => 'INTEGER',
+        ])->execute();
+        $this->db->createCommand()->insert($this->migrationTable, [
+            'version' => self::BASE_MIGRATION,
+            'apply_time' => time(),
+        ])->execute();
+        $this->stdout("Done.\n", Console::FG_GREEN);
+    }
+}